backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.0
google-cloud-aiplatform==1.72.0
firebase-admin==6.6.0
python-multipart==0.0.9
PyMuPDF==1.25.0
python-docx==1.1.0
langchain-text-splitters==0.3.0
pydantic-settings==2.5.0
